fix: share single CockroachDB testcontainer across reconciliation persistence tests
Each test was spinning up its own CockroachDB container (~12s each),
causing the 51-test suite to exceed the default 10-minute timeout.
Introduces TestMain to create one shared container with table truncation
between tests for data isolation. Reduces suite runtime from 600s+ to ~14s.
